# repo-root/groups/group-01-manage-field/openapi.yaml
openapi: 3.0.3
info:
  title: Field Management API
  version: 2.0.0
  description: HTTP contract for fields, field statuses, reservations, and field availability.
servers:
  - url: http://localhost:8080
    description: Base URL

tags:
  - name: Reference Data
  - name: Fields
  - name: Reservations

paths:
  /v1/field-statuses:
    get:
      tags: [Reference Data]
      summary: List field statuses
      operationId: listFieldStatuses
      responses:
        '200':
          description: Field status list
          content:
            application/json:
              schema:
                $ref: './tests/schemas/field-status-list.schema.json'

  /v1/reservation-statuses:
    get:
      tags: [Reference Data]
      summary: List reservation statuses
      operationId: listReservationStatuses
      responses:
        '200':
          description: Reservation status list
          content:
            application/json:
              schema:
                $ref: './tests/schemas/reservation-status-list.schema.json'

  /v1/fields/available:
    get:
      tags: [Fields]
      summary: List available fields
      description: Return active fields that are not blocked by a pending or confirmed reservation on the requested slot. The slot must stay on a single date, start on a full hour or half hour, and last 60, 90, or 120 minutes.
      operationId: listAvailableFields
      parameters:
        - $ref: '#/components/parameters/ReservationDateQueryParam'
        - $ref: '#/components/parameters/ReservationStartTimeQueryParam'
        - $ref: '#/components/parameters/ReservationEndTimeQueryParam'
      responses:
        '200':
          description: Available field list
          content:
            application/json:
              schema:
                $ref: './tests/schemas/field-list.schema.json'
        '400':
          $ref: '#/components/responses/BadRequest'

  /v1/fields/{field_id}:
    parameters:
      - $ref: '#/components/parameters/FieldIdParam'
    get:
      tags: [Fields]
      summary: Get a field
      description: Return a field with its status and the reservations attached to it, including each reservation status.
      operationId: getField
      responses:
        '200':
          description: Field detail
          content:
            application/json:
              schema:
                $ref: './tests/schemas/field-detail.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/fields/{field_id}/status:
    parameters:
      - $ref: '#/components/parameters/FieldIdParam'
    patch:
      tags: [Fields]
      summary: Change a field status
      operationId: updateFieldStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFieldStatusRequest'
      responses:
        '200':
          description: Updated field
          content:
            application/json:
              schema:
                $ref: './tests/schemas/field.schema.json'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/fields/{field_id}/reservations:
    parameters:
      - $ref: '#/components/parameters/FieldIdParam'
    get:
      tags: [Reservations]
      summary: List field reservations
      description: Return all reservations attached to one field, including their reservation statuses.
      operationId: listFieldReservations
      responses:
        '200':
          description: Field reservation list
          content:
            application/json:
              schema:
                $ref: './tests/schemas/reservation-list.schema.json'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Reservations]
      summary: Create a reservation
      description: Primary synchronous contract used by Match Management to reserve a field on a same-day slot aligned on half-hour boundaries.
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: './tests/schemas/reservation.schema.json'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/fields/{field_id}/reservations/{reservation_id}/status:
    parameters:
      - $ref: '#/components/parameters/FieldIdParam'
      - $ref: '#/components/parameters/ReservationIdParam'
    patch:
      tags: [Reservations]
      summary: Change a reservation status
      description: Update the status of an existing reservation.
      operationId: updateReservationStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationStatusRequest'
      responses:
        '200':
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: './tests/schemas/reservation.schema.json'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    FieldIdParam:
      name: field_id
      in: path
      required: true
      schema:
        $ref: './tests/schemas/id.schema.json#/properties/id'
    ReservationIdParam:
      name: reservation_id
      in: path
      required: true
      schema:
        $ref: './tests/schemas/id.schema.json#/properties/id'
    ReservationDateQueryParam:
      name: date
      in: query
      required: true
      description: Same-day reservation date in `YYYY-MM-DD` format.
      schema:
        $ref: './tests/schemas/time-slot.schema.json#/definitions/date'
    ReservationStartTimeQueryParam:
      name: start_time
      in: query
      required: true
      description: Reservation start time in `HH:MM` format on a full hour or half hour.
      schema:
        $ref: './tests/schemas/time-slot.schema.json#/definitions/start_time'
    ReservationEndTimeQueryParam:
      name: end_time
      in: query
      required: true
      description: Reservation end time in `HH:MM` format on a full hour or half hour.
      schema:
        $ref: './tests/schemas/time-slot.schema.json#/definitions/end_time'

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '../../openapi/common/schemas.yaml#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '../../openapi/common/schemas.yaml#/components/schemas/ErrorResponse'
    Conflict:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '../../openapi/common/schemas.yaml#/components/schemas/ErrorResponse'

  schemas:
    UpdateFieldStatusRequest:
      type: object
      properties:
        status_id:
          $ref: './tests/schemas/id.schema.json#/properties/id'
      required:
        - status_id
      additionalProperties: false

    CreateReservationRequest:
      $ref: './tests/schemas/create-reservation-request.schema.json'

    UpdateReservationStatusRequest:
      type: object
      properties:
        status_id:
          $ref: './tests/schemas/id.schema.json#/properties/id'
      required:
        - status_id
      additionalProperties: false

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY

security:
  - ApiKeyAuth: []